# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ConceptNodeView": {
        "description": "Public concept DTO: citeable metadata without internal embedding vectors.",
        "properties": {
          "aliases": {
            "items": {
              "type": "string"
            },
            "title": "Aliases",
            "type": "array"
          },
          "applications": {
            "items": {
              "type": "string"
            },
            "title": "Applications",
            "type": "array"
          },
          "canonical_name": {
            "title": "Canonical Name",
            "type": "string"
          },
          "concept_id": {
            "title": "Concept Id",
            "type": "string"
          },
          "definition": {
            "default": "",
            "title": "Definition",
            "type": "string"
          },
          "evidence_refs": {
            "items": {
              "$ref": "#/components/schemas/EvidenceRef"
            },
            "title": "Evidence Refs",
            "type": "array"
          },
          "graph_metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Graph Metrics",
            "type": "object"
          },
          "importance_score": {
            "default": 0.0,
            "title": "Importance Score",
            "type": "number"
          },
          "key_points": {
            "items": {
              "type": "string"
            },
            "title": "Key Points",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "prerequisites": {
            "items": {
              "type": "string"
            },
            "title": "Prerequisites",
            "type": "array"
          },
          "source_count": {
            "default": 0,
            "title": "Source Count",
            "type": "integer"
          },
          "summary": {
            "default": "",
            "title": "Summary",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "concept_id",
          "name",
          "canonical_name"
        ],
        "title": "ConceptNodeView",
        "type": "object"
      },
      "CourseGraphMeta": {
        "description": "总图谱嵌套结构元数据，仅总图谱有值。",
        "properties": {
          "children_map": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Children Map",
            "type": "object"
          },
          "core_concept_ids": {
            "items": {
              "type": "string"
            },
            "title": "Core Concept Ids",
            "type": "array"
          },
          "source_session_ids": {
            "items": {
              "type": "string"
            },
            "title": "Source Session Ids",
            "type": "array"
          }
        },
        "title": "CourseGraphMeta",
        "type": "object"
      },
...
      "EdgeType": {
        "enum": [
          "MENTIONS",
          "RELATES_TO",
          "CO_OCCURS_WITH",
          "CONTAINS"
        ],
        "title": "EdgeType",
        "type": "string"
      },
      "EvidenceRef": {
        "properties": {
          "chunk_id": {
            "title": "Chunk Id",
            "type": "string"
          },
          "locator": {
            "title": "Locator",
            "type": "string"
          },
          "score": {
            "default": 0.0,
            "title": "Score",
            "type": "number"
          },
          "snippet": {
            "title": "Snippet",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_type": {
            "$ref": "#/components/schemas/SourceKind"
          }
        },
        "required": [
          "chunk_id",
          "source_id",
          "source_type",
          "locator",
          "snippet"
        ],
        "title": "EvidenceRef",
        "type": "object"
      },
...
      "GlobalConceptHit": {
        "properties": {
          "canonical_name": {
            "title": "Canonical Name",
            "type": "string"
          },
          "concept_id": {
            "title": "Concept Id",
            "type": "string"
          },
          "course_title": {
            "title": "Course Title",
            "type": "string"
          },
          "importance_score": {
            "title": "Importance Score",
            "type": "number"
          },
          "lecture_title": {
            "title": "Lecture Title",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "course_title",
          "lecture_title",
          "concept_id",
          "name",
          "canonical_name",
          "importance_score"
        ],
        "title": "GlobalConceptHit",
        "type": "object"
      },
      "GraphArtifactView": {
        "properties": {
          "built_at": {
            "format": "date-time",
            "title": "Built At",
            "type": "string"
          },
          "concepts": {
            "items": {
              "$ref": "#/components/schemas/ConceptNodeView"
            },
            "title": "Concepts",
            "type": "array"
          },
          "course_meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CourseGraphMeta"
              },
              {
                "type": "null"
              }
            ]
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "title": "Edges",
            "type": "array"
          },
          "schema_version": {
            "default": "2.0",
            "title": "Schema Version",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "topic_clusters": {
            "items": {
              "$ref": "#/components/schemas/TopicClusterNode"
            },
            "title": "Topic Clusters",
            "type": "array"
          }
        },
        "required": [
          "session_id",
          "built_at"
        ],
        "title": "GraphArtifactView",
        "type": "object"
      },
      "GraphEdge": {
        "properties": {
          "edge_id": {
            "title": "Edge Id",
            "type": "string"
          },
          "edge_type": {
            "$ref": "#/components/schemas/EdgeType"
          },
          "properties": {
            "additionalProperties": true,
            "title": "Properties",
            "type": "object"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "target": {
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "source",
          "target",
          "edge_type"
        ],
        "title": "GraphEdge",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NodeType": {
        "enum": [
          "concept",
          "topic_cluster"
        ],
        "title": "NodeType",
        "type": "string"
      },
...
      "SearchChunkHit": {
        "properties": {
          "chunk_id": {
            "title": "Chunk Id",
            "type": "string"
          },
          "page_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page End"
          },
          "page_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Start"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_type": {
            "$ref": "#/components/schemas/SourceKind"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "time_end": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time End"
          },
          "time_start": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Start"
          }
        },
        "required": [
          "chunk_id",
          "source_id",
          "source_type",
          "score",
          "text"
        ],
        "title": "SearchChunkHit",
        "type": "object"
      },
      "SearchConceptHit": {
        "properties": {
          "canonical_name": {
            "title": "Canonical Name",
            "type": "string"
          },
          "concept_id": {
            "title": "Concept Id",
            "type": "string"
          },
          "evidence_chunk_ids": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Chunk Ids",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "source_count": {
            "default": 0,
            "title": "Source Count",
            "type": "integer"
          }
        },
        "required": [
          "concept_id",
          "name",
          "canonical_name",
          "score"
        ],
        "title": "SearchConceptHit",
        "type": "object"
      },
      "SearchRequest": {
        "properties": {
          "limit": {
            "default": 8,
            "title": "Limit",
            "type": "integer"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "query"
        ],
        "title": "SearchRequest",
        "type": "object"
      },
      "SearchResponse": {
        "properties": {
          "chunks": {
            "items": {
              "$ref": "#/components/schemas/SearchChunkHit"
            },
            "title": "Chunks",
            "type": "array"
          },
          "concepts": {
            "items": {
              "$ref": "#/components/schemas/SearchConceptHit"
            },
            "title": "Concepts",
            "type": "array"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "query"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
...
      "SourceKind": {
        "enum": [
          "pdf",
          "audio",
          "document",
          "image",
          "video"
        ],
        "title": "SourceKind",
        "type": "string"
      },
      "SubgraphEdge": {
        "properties": {
          "edge_type": {
            "$ref": "#/components/schemas/EdgeType"
          },
          "properties": {
            "additionalProperties": true,
            "title": "Properties",
            "type": "object"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "target": {
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "source",
          "target",
          "edge_type"
        ],
        "title": "SubgraphEdge",
        "type": "object"
      },
      "SubgraphNode": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "node_type": {
            "$ref": "#/components/schemas/NodeType"
          }
        },
        "required": [
          "id",
          "label",
          "node_type"
        ],
        "title": "SubgraphNode",
        "type": "object"
      },
      "SubgraphResponse": {
        "properties": {
          "center_concept_id": {
            "title": "Center Concept Id",
            "type": "string"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/SubgraphEdge"
            },
            "title": "Edges",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/SubgraphNode"
            },
            "title": "Nodes",
            "type": "array"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "center_concept_id"
        ],
        "title": "SubgraphResponse",
        "type": "object"
      },
...
      "TopicClusterNode": {
        "properties": {
          "cluster_id": {
            "title": "Cluster Id",
            "type": "string"
          },
          "concept_ids": {
            "items": {
              "type": "string"
            },
            "title": "Concept Ids",
            "type": "array"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "cluster_id",
          "title",
          "summary"
        ],
        "title": "TopicClusterNode",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Corpus2Node API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/graph/concepts": {
      "get": {
        "description": "Substring-search concepts across every built graph (for the global search bars).",
        "operationId": "search_concepts_global_api_graph_concepts_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GlobalConceptHit"
                  },
                  "title": "Response Search Concepts Global Api Graph Concepts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Concepts Global",
        "tags": [
          "graph"
        ]
      }
    },
    "/api/graph/search": {
      "post": {
        "operationId": "search_graph_api_graph_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Graph",
        "tags": [
          "graph"
        ]
      }
    },
    "/api/graph/{session_id}": {
      "get": {
        "operationId": "get_graph_api_graph__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphArtifactView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Graph",
        "tags": [
          "graph"
        ]
      }
    },
    "/api/graph/{session_id}/subgraph": {
      "get": {
        "operationId": "get_subgraph_api_graph__session_id__subgraph_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "concept_id",
            "required": true,
            "schema": {
              "title": "Concept Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "depth",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Depth",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "max_nodes",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Max Nodes",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubgraphResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Subgraph",
        "tags": [
          "graph"
        ]
      }
    },
...
    "/graph/concepts": {
      "get": {
        "description": "Substring-search concepts across every built graph (for the global search bars).",
        "operationId": "search_concepts_global_graph_concepts_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GlobalConceptHit"
                  },
                  "title": "Response Search Concepts Global Graph Concepts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Concepts Global",
        "tags": [
          "graph"
        ]
      }
    },
    "/graph/search": {
      "post": {
        "operationId": "search_graph_graph_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Graph",
        "tags": [
          "graph"
        ]
      }
    },
    "/graph/{session_id}": {
      "get": {
        "operationId": "get_graph_graph__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphArtifactView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Graph",
        "tags": [
          "graph"
        ]
      }
    },
    "/graph/{session_id}/subgraph": {
      "get": {
        "operationId": "get_subgraph_graph__session_id__subgraph_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "concept_id",
            "required": true,
            "schema": {
              "title": "Concept Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "depth",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Depth",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "max_nodes",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Max Nodes",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubgraphResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Subgraph",
        "tags": [
          "graph"
        ]
      }
    },
...
  }
}